Fix bugs with writing /data and /sdcard.
/data and /sdcard are managed as a combination. Manage information of /data maybe writen to the area of /sdcard vice versa.
'fastboot -w flashall' will flash boot.img recovery.img and syste.img and erase /data and /cache.
If manager information of /sdcard is writen to the area of /data, erasing of /data will destrory them. It will lead fatal errors.
If manager information of /data is writen to the area of /sdcard, erasing of /data can not clean up /data at all. It also will lead fatal errors.
The error is like this:
ERROR:ret shouldn't be -6!!!!